Add Company class common getDescription method
The Company class provides the ability to list all people associated
with the Company. It stores instances of Person in an ArrayList. This
includes shareholders who may or may not be Employees.

Add getDescription method to Person and all child classes. This
provides a string description for output in Company class.

Add getShare and setShare methods to Person class. Method setShare
takes Share instance as a parameter. This is in line with other methods
and constructors as these can be created dynamically.

Change Share fields to be private. They should have really been
private in the first place.

Closes-Todo: #2
Commit: #14