Add calculatePayment method to Company class
Use for each loop to return name, dividend due, employment number and
payment number. Print each out with an identifying label before. Seems
sensible to define this functionality in the Company class rather than
in each individual subclass of person, but essentially the same code
would be used (ie. if a calculatePayment method was made in each
class).

Closes-Todo: #4
Commit: #15